+++ Js2IL.Tests/Node/JavaScript/SetImmediate_Nested_ExecutesInNextIteration.js
@@ -1,0 +1,6 @@
+setImmediate(() => {
+    console.log("outer");
+    setImmediate(() => console.log("inner"));
+});
+
+console.log("scheduled nested immediates.");